requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy
asyncpg
pydantic>=2.3.0
bcrypt==3.2.0
passlib[bcrypt]
python-jose[cryptography]>=3.3.0